Ignore the revert commit, I just messed up a bit.

+++ src/main/java/se/chalmers/bokforing/model/Sum.java
@@ -24,12 +24,12 @@ public class Sum implements Serializable {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private int id;
     
-    private int value;
+    private double value;
     
     @Enumerated(EnumType.STRING)
     private PostType type;
     
-    public Sum(int value, PostType type){
+    public Sum(double value, PostType type){
         this.value = value;
         this.type = type;
     }
@@ -37,14 +37,14 @@ public class Sum implements Serializable {
     /**
      * @return the value
      */
-    public int getValue() {
+    public double getValue() {
         return value;
     }
 
     /**
      * @param value the value to set
      */
-    public void setValue(int value) {
+    public void setValue(double value) {
         this.value = value;
     }
